refactor: reorganize shared functions
