+++ src/core/ts-type/checker/is-array.ts
