+++ src/core/ts-type/checker/is-class.ts
