+++ src/core/ts-type/checker/is-extends-target-super-class.ts
