+++ src/core/ts-type/checker/private/get-symbol.ts
