+++ src/core/ts-type/finder/array-element.ts
@@ -1,6 +1,7 @@
 import { Type } from "typescript";
 
-import { isArrayType } from "../type-checker/is-array";
+import { isArrayType } from "../checker/is-array";
+
 
 /**
  * Find the element type of an array type (e.g. for s3.Bucket[], returns s3.Bucket)
