+++ src/core/ts-type/finder/constructor-property-name.ts
