+++ src/rules/construct-constructor-property.ts
@@ -6,8 +6,8 @@ import {
   TSESTree,
 } from "@typescript-eslint/utils";
 
+import { findConstructor } from "../core/ast-node/finder/constructor";
 import { isConstructType } from "../core/cdk-construct/type-checker/is-construct";
-import { findConstructor } from "../shared/ast-node-finder/constructor";
 import { createRule } from "../shared/create-rule";
 
 type Context = TSESLint.RuleContext<
