+++ src/rules/no-unused-props/index.ts
@@ -7,8 +7,8 @@ import {
 } from "@typescript-eslint/utils";
 import { Type } from "typescript";
 
+import { findConstructor } from "../../core/ast-node/finder/constructor";
 import { isConstructType } from "../../core/cdk-construct/type-checker/is-construct";
-import { findConstructor } from "../../shared/ast-node-finder/constructor";
 import { createRule } from "../../shared/create-rule";
 
 import { PropsUsageAnalyzer } from "./props-usage-analyzer";
